+++ kurento-integration-tests/kurento-test/src/test/java/org/kurento/test/functional/recorder/RecorderPlayerDisconnectTest.java
@@ -192,7 +192,7 @@ public class RecorderPlayerDisconnectTest extends BaseRecorder {
 
     AssertMedia.assertCodecs(recordingFile, expectedVideoCodec, expectedAudioCodec);
     AssertMedia.assertDuration(recordingFile, TimeUnit.SECONDS.toMillis(playtime),
-        TimeUnit.SECONDS.toMillis(getPage().getThresholdTime()));
+        TimeUnit.SECONDS.toMillis((long)getPage().getThresholdTime()));
 
     // Release Media Pipeline #2
     mp2.release();
